backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.2
langgraph>=0.0.20
pinecone
snowflake-connector-python>=3.0.0
pandas>=2.0.0
matplotlib>=3.7.0
tavily-python>=0.2.2
numpy>=1.24.0
requests>=2.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
sentence-transformers>=2.2.2